+++ Fossology.Rest.Dotnet.Model/Group.cs
@@ -1,0 +1,42 @@
+﻿// ---------------------------------------------------------------------------
+// <copyright file="Group.cs" company="Tethys">
+//   Copyright (C) 2022 T. Graf
+// </copyright>
+//
+// Licensed under the MIT License.
+// SPDX-License-Identifier: MIT
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
+// either express or implied.
+// ---------------------------------------------------------------------------
+
+namespace Fossology.Rest.Dotnet.Model
+{
+    using Newtonsoft.Json;
+
+    /// <summary>
+    /// Fossology Group information.
+    /// </summary>
+    public class Group
+    {
+        /// <summary>
+        /// Gets or sets the id.
+        /// </summary>
+        [JsonProperty("id")]
+        public int Id { get; set; }
+
+        /// <summary>
+        /// Gets or sets the name.
+        /// </summary>
+        [JsonProperty("name")]
+        public string Name { get; set; }
+
+        /// <inheritdoc />
+        public override string ToString()
+        {
+            return $"{this.Id}: {this.Name}";
+        }
+    } // Group
+}