+++ ChangeLog.md
@@ -5,6 +5,7 @@
 * supports FOSSology REST API v1.4.3 (FOSSology 4.0.0 built @ 2023/03/09).
 * new method GetHealth().
 * have `FossologyClient` to be a partial class for better overview.
+* new method GetGroupList().
 
 ## 1.1.0 (2020-06-25)
 * supports FOSSology REST API v1.0.16 (FOSSology 3.8.0 built @ 2020/06/19).